requirements.txt
# Core
python-dotenv==1.0.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# LLM
openai==1.3.7
tiktoken==0.5.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2
dateparser==1.2.0
pytz==2023.3

# CLI
click==8.1.7
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
mypy==1.7.1
